Split ev/core/commands.py into a package by domain
Same extract-and-recompose approach as the tools.py and memory.py
splits: Commands' ~70 deterministic slash commands become mixins
composed in commands/base.py (which also keeps the dispatch table,
help(), run()/runnable(), and the shared _pick()/_now()/_month_bounds()
helpers), with commands/__init__.py re-exporting Commands and
COMMAND_LIST so every existing import path and call site keeps
working unchanged. No behavior change.
